chore: refactor code to keep convention
